+++ app/src/main/java/de/welthungerhilfe/cgm/scanner/datasource/models/ResultChildDistance.java
@@ -20,13 +20,13 @@ public class ResultChildDistance extends Results{
 
     public static class Data {
         @Expose
-        float child_distance;
+        double child_distance;
 
-        public float getChild_distance() {
+        public double getChild_distance() {
             return child_distance;
         }
 
-        public void setChild_distance(float child_distance) {
+        public void setChild_distance(double child_distance) {
             this.child_distance = child_distance;
         }
     }
