+++ app/src/main/java/de/welthungerhilfe/cgm/scanner/ui/activities/ScanModeActivity.java
@@ -755,7 +755,7 @@ public class ScanModeActivity extends BaseActivity implements View.OnClickListen
         if (mIsRecording && (frameIndex % AppConstants.SCAN_FRAMESKIP == 0)) {
 
             float light = mCameraInstance.getLightIntensity();
-            float child_distance = mCameraInstance.getTargetDistance();
+            double child_distance = mCameraInstance.getTargetDistance();
             if (light > 1) {
                 light = 1.0f - (light - 1.0f);
             }
@@ -881,7 +881,7 @@ public class ScanModeActivity extends BaseActivity implements View.OnClickListen
         }
     }
 
-    private void onProcessArtifact(File artifactFile, ArtifactType type, float childHeight, float poseScore, String poseCordinates,float child_distance, float light_score) {
+    private void onProcessArtifact(File artifactFile, ArtifactType type, float childHeight, float poseScore, String poseCordinates,double child_distance, float light_score) {
         if (artifactFile.exists()) {
             FileLog log = new FileLog();
